Use Nexus style data layout, allow data components on external sdcard and some optimizations and cleanups
Use internal data as in Nexus devices (sdcard0 is emulated) and external sdcard now called sdcard1.

Optional second partition of external sdcard (data1) can be used to store selected data components such as:
app, data, media and dalvik-cache. To do so, one must create a folder with a name of the component/s on data1.

We register sdcard1 as secondary storage allowing built-in asec mechanism to be used to move apps to it.

Also:
* Add misc recovery related settings and coresponding changes to TWRP in order to support above.
* Modify recovery to backup of emulated internal sdcard and apps in asac container.
* Remove duplicated default init.rc code
* Optimize storage, use cache as secondary swap and reuse devlog as cache

To summarize, this patch allows user to encrypt device including sdcard0,
place selected data components on external sdcard and hopefuly provides better performance.

Change-Id: Ib52086347eb1f8fc9fd11031428f2c49bdda9afb